Removed local_interfaces configuration directory. Instead, the user_overrides files for an
interface may list and number of 'feeds' (local files only at present).
Each feed is an interface file, which is read and merged with the upstream one.
For example, ROX-Filer CVS contains a ROX-Filer.xml interface listing just that version.
You can then register this feed with the injector to get the CVS version listed
as a choice along with the upstream versions.

git-svn-id: file:///home/talex/Backups/sf.net/Subversion/zero-install/injector/injector@436 9f8c893c-44ee-0310-b757-c8ca8341c71e