Moved autopolicy.NeedDownload into zeroinstall package (old references will
still work) and made it a subclass of SafeException.
When a feed can't be fetched due to NeedDownload, rethrow the exception rather
than logging a warning, so that it will get downloaded.

Updated 0launch --add-feed to new system, and added support for removing
feeds; the option is now called --feed to reflect this change.

git-svn-id: file:///home/talex/Backups/sf.net/Subversion/zero-install/injector/injector@565 9f8c893c-44ee-0310-b757-c8ca8341c71e