Added support for XML comment signatures. These are specially formatted
comment blocks at the end of the XML containing a base64 encoded string
which is a GPG detached signature.

The base64 encoding ensures that people can't sneak XML special characters
into the signature (which would be ignored by GPG but might confuse XML
parsers).

The result of vertifying an XML-signed file is the original data (with the
signature), not the data alone (as it is with the old signatures). This is
possible because the signed version is now valid XML.

git-svn-id: file:///home/talex/Backups/sf.net/Subversion/zero-install/injector/injector@621 9f8c893c-44ee-0310-b757-c8ca8341c71e