Moved 0launch-gui inside 0launch/zeroinstall. Adjusted setup.py so that a copy of the GUI
will be included in each release of the injector. This bundled copy appears as an additional
(local) feed. The idea is that the injector won't need to fallback to console mode on the
first run. Because the local copy is only a feed, the injector will still check for and
use updates, as usual.

git-svn-id: file:///home/talex/Backups/sf.net/Subversion/zero-install/trunk/0launch@661 9f8c893c-44ee-0310-b757-c8ca8341c71e