Instead of always adding implementations to the user's store, try writing
to the first of the system stores. Only use the user store if unsuccessful.
This makes it easy for root to install things for users (at least, until we
have the set-uid helper). The system store will be created world-readable,
and anything stored in it will be read-only. User stores honour the user's
umask as usual.

git-svn-id: file:///home/talex/Backups/sf.net/Subversion/zero-install/trunk/0launch@1463 9f8c893c-44ee-0310-b757-c8ca8341c71e