Use iface_cache.get_interface() rather than policy.get_interface() in several
places where we don't expect to trigger a download. The idea is that the
policy version should become private.

For present-but-stale interfaces, Policy.get_interface() just adds it to a
new stale_feeds field rather than starting the download. recalculate() checks
this at the end and starts them all off. The plan is to make this step
optional, so that we can tell the difference between stale and missing
interfaces easily.

git-svn-id: file:///home/talex/Backups/sf.net/Subversion/zero-install/trunk/0launch@1480 9f8c893c-44ee-0310-b757-c8ca8341c71e