Added --manpage option to 0alias. This displays the man-page for an alias, e.g.
$ 0alias foo http://example.com/foo.xml
$ 0alias --manpage foo

If the argument doesn't look like a 0alias script, it is passed to the system's
man command, so you can do this in your .bashrc file:

alias man=’0alias --manpage’

And then do:

$ 0alias foo http://example.com/foo.xml
$ man foo

Currently, the man-page is located in the implementation using some ugly heuristics. Ideally,
the feed should say exactly the man-page is.

git-svn-id: file:///home/talex/Backups/sf.net/Subversion/zero-install/trunk/0launch@1734 9f8c893c-44ee-0310-b757-c8ca8341c71e