Change how last-check_attempt is stored.
Instead of storing this as an attribute in the user_overrides document for a feed,
use an empty timestamp file. The time is now read from iface_cache, not stored
as a property on the Interface / ZeroInstallFeed, as it wasn't really a property of
either (it exists before the feeds does, but is a per-feed value, not per-interface).