When looking up key owners, cope with subordinate keys.
The injector always records the fingerprint of the actual signing key in the
trust database, which may be a subordinate key. However, when looking up the
user information (for the Preferences box) it only looked at the fingerprints
of the primary keys.

This meant that it was unable to display the user details for some trusted
keys, and printed "Got information about key ..., but I only asked about ...!"
warnings.