Change to mirror logic; keep waiting for primary
In the previous code, if the upstream site didn't reply before the mirror then
we stopped waiting for it. With a fast mirror and a slow upstream, you might
never see the (newer) upstream version.

Now, we start a mirror download if the primary fails or takes too long (as
before), but we don't abort the primary if the mirror succeeds. We do abort the
mirror if the primary succeeeds, though.