Replaced "Interface Properties" button with a menu button on each row of the table.
This should make the other options (e.g. Compile and Report Bug) easier to find.
The Refresh and Show Cache buttons move down into the new space, making the window shorter
overall.

Also, elide the end of the summary when the window is too narrow, rather than showing a
scrollbar.