Added Implementation.local_path attribute
This is the absolute path of the implementation for local implementations, or
None if it's not local (a distribution package or a Zero Install package
identified by a digest, whether cached or not).

This is so that we can avoid id.startswith('/') checks everywhere, since this
doesn't work on Windows.