Select best version one interface at a time
Instead of minimising a global cost function over all components at once, just
try to select the locally best implementation of the root interface. Then
select the best version of the first dependency, etc.

Note that we still apply all *constraints* globally (we choose a version of the
root that is compatible with some combination of the other components), but of
the legal combinations, we decide on the best one using only local information.

For example, we can no longer select an older version of the application in
order to get later versions of several dependencies. This is much faster, and
easier to understand.