If we fail to find a set of versions, try to return a close match
After failing to solve, try again with an extra dummy version for every interface.
This matches any restriction and any architecture and has no dependencies, but we
only select it if we have no other choice.

The idea is that failed selections give the user a better idea of where the
problem is.