API changes: use Feed rather than Interface in many places
Originally, we only had Interfaces, which included the XML. Then we split this
into Interface for configuration (e.g. user's preferred stability policy) and
Feeds (for the XML).

Before, iface_cache only cached interfaces (each of which contained a feed).
That meant that if you only wanted a feed, we had to make up a dummy interface
for it. This patch updates the terminology in many places and passes around
Feed objects where it makes sense.

To see if you're using a deprecated API, uncomment the "warnings" lines in
model.py.