Model information about distribution packages as separate feeds
Before, reading in a feed containing a <package-implementation> element
would expand this to a list of packages provided by the distribution's
package manager (installed or not) and include them in the feeds
'implementations'.

Now, feed with such an element has a corresponding "distribution feed", which
is the feed's URL prefixed with "distribuion:". This feed contains the
distribution packages.

Initially, it contains only installed packages. After updating it, it may
also contain candidate (uninstalled) version.

This has two advantages:

- The loaded representation of feed no longer depends on the host
  distribution, which is cleaner.
- Fetching candidates can be done asynchronously.