Added support for $-substitution in <arg> elements
This is to support programs that need to be given their dependencies
through arguments. For example:

    <command name="run" path="myprog">
      <arg>--java-home=${JAVA_HOME}</arg>
    </command>

Currently, the values come from the environment (so any environment variable
can be used). Possibly it should be restricted to only those variables set in
the feed, or extended to other types of binding (to avoid polluting the
process's environment just to pass an argument).