Fixed updating of last-check-attempt
We were only updating this during background updates, and only for feeds of the
selected interfaces. However, the solve might need to look at feeds for
interfaces that aren't used.

For example, older versions of ROX-Filer depend on libglade. The libglade feed
would then never get marked as checking, and we would keep retrying if the
network was unavailable.