Fixed bug when a feed is used for multiple interfaces
If we rename an interface 'old' to 'new', we may want to import the contents of
'old' into 'new' using <feed>. If another feed depends on both interfaces (e.g.
version 1 depends on 'old' and version 2 depends on 'new') then there is a
problem: each <implementation> in 'old' belongs to two interfaces. This resulted
in a crash:

  File "../zeroinstall/injector/solver.py", line 466, in add_iface
    assert impl not in impl_to_var
  AssertionError