Cope better if no GUI is available
We used to assume that a GUI was available whenever $DISPLAY was set. However,
there could be other reasons why the GUI can't be used, such as missing Python
modules. We now detect this and fall back to console mode gracefully.

If --gui is specified (to force use of the GUI) and no GUI is available, we
display an error message and abort, as before.