"0install select" now always does a fresh selection
Before, if it was used with an app and no additional requirements, it just
showed the current selections. But now, use "0install show" for that.

For apps, it now shows what changed too, and prompts you to use "0install
update" to save the changes if either the requirements or the selections
changed (before, it only prompted if the requirements had changed).