Added an explicit dependency on python-gobject
This is needed on Fedora to prevent us running 0install with Python 3 when
we only have the Python 2 version of python-gobject.

It will also be needed on OS X to make sure we don't mix the native version
of Python when only the MacPorts Python has python-gobject.

Requested by Tim Cuthbertson.