Added Scope_filter.may_compile
When may_compile is set, we make a fake binary for each source impl,
representing the binary that could be produced by compiling it.

We use the same logic that 0compile uses to generate the binary XML.
In future, we should be able to get 0compile to use this code directly.

Based on Tim Cuthbertson's autocompile patches. However, we do not
include source dependencies when solving (we just assume that any source
implementation can be compiled). This matches what 0compile's
autocompile feature does, and avoids pulling in large numbers of extra
feeds just in case we need to compile.