Make packagekit an explicit argument to distributions
Avoids needing a global mutable ref for the unit-tests.

One problem is that obus grabs the environment variables the first time
it needs them. If anything initialises PackageKit before the obus test,
that test will fail (and we can never do two obus-using tests, ever).
Therefore, we avoid trying to initialise obus if the environment
indicates that we're running unit-tests.